Grab a token to prepare for hitting the JSON API
I've discovered this is about as far as we can get by scraping the page
with Mechanize, since it doesn't support JavaScript. We'll need to
figure out which XHR calls Bandcamp is making to get further data, and
make them ourselves.

My first discovery is that every request requires a token from the last
request, and helpfully we can grab a token from the DOM for that.