Keep track of tralbum_type for each album/track
Turns out setting "tralbum_type" to "a" in every request here doesn't
work - Bandcamp expects an accurate value 😄

Seems like "a" means album and "t" means track. There are probably other
types too. We now grab this when we source albums, and pass it along
when we fetch their collectors.