Make albums more look-up-able
This stores albums in the same Hash format I used for collectors, with
the `tralbum_id` as the key. Each album also now keeps a list of
`collector_ids`, so at the end I can hopefully find the albums with the
most collectors that doesn't include me. Those'll be albums I probably
want to check out.